refactor: antitheft and message validation to modules
+ moved antitheft from data -> modules
+ moved validateMessage from data -> modules
+ renamed validateMessage to validateFrameMessage to start alignment w/ frame.js API
